user controller finalization

+++ app/Http/Controllers/UserController.php
@@ -18,16 +18,16 @@ class UserController extends Controller
         //
     }
 
-   public function index(Request $request)
+   public function getUserAll(Request $request)
     {
         //show all users
         $user = User::all();
 
         try {
-            if ($user) {
+            if ($request->auth->role == 'admin') {
                 return response()->json([
                     'success' => true,
-                    'message' => 'User ditemukan',
+                    'message' => 'User found',
                     'data' => [
                         'user' => $user
                     ]
@@ -35,8 +35,8 @@ class UserController extends Controller
             } else {
                 return response()->json([
                     'success' => false,
-                    'message' => 'Gagal mengambil data user'
-                ], 400);
+                    'message' => 'Data restricted'
+                ], 403);
             }
         } catch (\Exception $e) {
             return response()->json([
@@ -46,34 +46,48 @@ class UserController extends Controller
         }
     }
 
-    public function getUserById(Request $request, $idUser)
+    public function getUserById(Request $request, $userId)
     {
-        $user = null;
-
-        /// Request user access if role == admin
-        $userRequest = $request->auth;
-        if ($userRequest == 'admin') {
-            $user = User::find($idUser);
-        } else {
-            $user = User::find($userRequest->id);
-        }
 
         // Get data user by ID
         try {
-            if ($user) {
-                return response()->json([
-                    'success' => true,
-                    'message' => 'User ditemukan',
-                    'data' => [
-                        'user' => $user
-                    ]
-                ], 200);
+            $userRequest = $request->auth;
+
+            if ($userRequest->role == 'user') {
+                // Check if user has access
+                if ($userRequest->id == $userId) {
+                    // return user data only
+                    return response()->json([
+                        'success' => true,
+                        'message' => 'User found',
+                        'data' => [
+                            'user' => User::find($userId)
+                        ]
+                    ], 200);
+                } else {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Data restricted'
+                    ], 403);
+                }
             } else {
-                return response()->json([
-                    'success' => false,
-                    'message' => 'Gagal mengambil data user'
-                ], 200);
-            }
+                $user = User::find($userId);
+
+                if ($user) {
+                    return response()->json([
+                        'success' => true,
+                        'message' => 'User found',
+                        'data' => [
+                            'user' => $user
+                        ]
+                    ], 200);
+                } else {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'User not found'
+                    ], 404);
+                }
+            }     
         } catch (\Exception $e) {
             return response()->json([
                 'success' => false,
@@ -82,52 +96,55 @@ class UserController extends Controller
         }
     }
 
-    public function updateUser(Request $request, $idUser)
+    public function updateUser(Request $request, $userId)
     {
-        $user = User::find($idUser);
+        $user = User::find($userId);
 
         // Update data by ID user
         try {
-            if ($user) {
+            if ($request->auth->role == 'user' && $request->auth->id == $userId) {
                 $user->name = $request->input('name');
                 $user->email = $request->input('email');
                 $user->password = Hash::make($request->input('password'));
                 $user->save();
                 return response()->json([
                     'success' => true,
-                    'message' => 'User berhasil diupdate'
+                    'message' => 'Update data user succeeded',
+                    'data' => [
+                        'user' => $user
+                    ]
                 ], 200);
             } else {
                 return response()->json([
                     'success' => false,
-                    'message' => 'ID User tidak ditemukan'
-                ], 404);
+                    'message' => 'Update not allowed'
+                ], 403);
             }
         } catch (\Exception $e) {
             return response()->json([
                 'success' => false,
                 'message' => 'Terjadi kesalahan pada server'
-            ]);
+            ], 500);
         }
     }
     
-    public function destroy(Request $request, $idUser)
+    public function destroy(Request $request, $userId)
     {
-        $user = User::find($idUser);
-
+        $user = User::find($userId);
+        
         // Delete data by ID user
         try {
-            if ($user) {
+            if ($request->auth->role == 'user' && $request->auth->id == $userId) {
                 $user->delete();
                 return response()->json([
                     'success' => true,
-                    'message' => 'User berhasil dihapus'
+                    'message' => 'User is deleted'
                 ], 200);
             } else {
                 return response()->json([
                     'success' => false,
-                    'message' => 'ID User tidak ditemukan'
-                ], 404);
+                    'message' => 'Forbidden data request'
+                ], 403);
             }
         } catch (\Exception $e) {
             return response()->json([
